Cast port value to unsigned short in CConfigLoader

+++ client/src/CConfigLoader.cpp
@@ -59,7 +59,7 @@ void CConfigLoader::Load()
 
 	GetPrivateProfileString(CONFIG_SECTION.c_str(), "nickname", "", CLocalPlayer::m_Name, sizeof(CLocalPlayer::m_Name), ms_sDataPath.c_str());
 	GetPrivateProfileString(CONFIG_SECTION.c_str(), "ip", "", CNetwork::m_IpAddress, 15, ms_sDataPath.c_str());
-	CNetwork::m_nPort = GetPrivateProfileInt(CONFIG_SECTION.c_str(), "port", Config::DEFAULT_PORT, ms_sDataPath.c_str());
+	CNetwork::m_nPort = (unsigned short)GetPrivateProfileInt(CONFIG_SECTION.c_str(), "port", Config::DEFAULT_PORT, ms_sDataPath.c_str());
 
 #if DEBUG
 	printf("%s %s %d\n", CLocalPlayer::m_Name, CNetwork::m_IpAddress, CNetwork::m_nPort);
